+++ crate_anon/crateweb/core/views.py
@@ -31,7 +31,7 @@ from django.http.request import HttpRequest
 from django.shortcuts import render
 from crate_anon.crateweb.core.utils import is_clinician, is_developer
 from crate_anon.crateweb.research.views import query_context
-from crate_anon.version import VERSION, VERSION_DATE
+from crate_anon.version import CRATE_VERSION, CRATE_VERSION_DATE
 
 log = logging.getLogger(__name__)
 
@@ -63,7 +63,7 @@ def home(request: HttpRequest) -> HttpResponse:
 
 def about(request: HttpRequest) -> HttpResponse:
     context = {
-        'VERSION': VERSION,
-        'VERSION_DATE': VERSION_DATE,
+        'VERSION': CRATE_VERSION,
+        'VERSION_DATE': CRATE_VERSION_DATE,
     }
     return render(request, 'about.html', context)