+++ tools/rebuild_reinstall.py
@@ -26,13 +26,13 @@
 
 import os
 import subprocess
-from crate_anon.version import VERSION  # , VERSION_DATE
+from crate_anon.version import CRATE_VERSION  # , VERSION_DATE
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
 SOURCE_ROOT = os.path.abspath(os.path.join(THIS_DIR, os.pardir))
 PACKAGE_DIR = os.path.join(SOURCE_ROOT, "built_packages")
 PACKAGE = 'crate'
-DEBVERSION = '{}-1'.format(VERSION)
+DEBVERSION = '{}-1'.format(CRATE_VERSION)
 PACKAGEFILE = os.path.join(
     PACKAGE_DIR,
     '{PACKAGE}_{DEBVERSION}_all.deb'.format(PACKAGE=PACKAGE,